// src/components/weather-details.tsx
import type { WeatherData } from '@/api/types';
import {format} from "date-fns";
import { Sunrise, Sunset, Compass, Gauge} from "lucide-react";
import { Card, CardContent, CardHeader, CardTitle } from './ui/card';


interface WeatherDetailsProps {
    data: WeatherData;    
}


const WeatherDetails = ({data}: WeatherDetailsProps) => {

    const { wind, main, sys } = data;   //sys is sunrise, sunset

    const getWindDirection = (degree:number)=>{
        const directions = ["N", "NE", "E", "SE","S", "SW", "W", "NW"];
      const index=
      Math.round(((degree %=360)< 0 ? degree + 360 : degree)/45) % 8;  //MORE THAN 360 like 361 module 1 <0 ? 361+361 : 361  if  772/45 either 360/45 
      return directions[index];
    }


    const formatTime = (timestamp: number) => {
return format(new Date(timestamp * 1000), "h:mm a")
    };

    const details = [
        {
            title: "Sunrise",
            value: formatTime(sys.sunrise),
            icon: Sunrise,
            color: "text-orange-500",
        },
        {
            title: "Sunset",
            value: formatTime(sys.sunset),
            icon: Sunset,
            color: "text-blue-500",
        },
        {
            title: "WInd Direction",
            value: `${getWindDirection(wind.deg)} (${wind.deg}°)`,
            icon: Compass,
            color: "text-green-500",
        },
        {
            title: "Pressure",
            value: `${main.pressure} hPa`,
            icon: Gauge,
            color: "text-purple-500",  
          },
    ];

  return(
  <Card>
    <CardHeader>
        <CardTitle>Card Title</CardTitle>
    </CardHeader>
    <CardContent>
<div className='grid gap-6 sm:grid-cols-2'>
    {details.map((detail)=>(
        <div key={detail.title}
         className='flex items-center gap-3 rounded-lg border p-4'>
            <detail.icon className={`h-5 w-5 ${detail.color}`} />
            <div>
                <p className='font-medium text-sm leading-none'>{detail.title}</p>
                <p className='text-sm text-muted-foreground'>{detail.value}</p>
            </div>
             </div>
    
    ))}
    </div>

   </CardContent>
  </Card>

  );
};

export default WeatherDetails
